optillm-desktop/backend/requirements.txt
fastapi
uvicorn[standard]
pyyaml
python-socketio
python-engineio
aiofiles
